src/api/requirements-cpu.txt
protobuf==3.20.1
huggingface_hub
llama-rs-python==0.0.2
openai
pynvml
farm-haystack[elasticsearch]
llm-rs==0.1.1

#api dependencies
psutil
fastapi
uvicorn
dependency_injector